♻️refactor: event.enum에 있던 일정과 할 일에 공통으로 쓰이는 enum 위치 변경
