+++ src/main/java/com/project/backend/domain/common/recurrence/enums/ExceptionType.java
@@ -1,4 +1,4 @@
-package com.project.backend.domain.event.enums;
+package com.project.backend.domain.common.recurrence.enums;
 
 public enum ExceptionType {
     SKIP, // 이 날짜는 발생하지 않음 (삭제)
