+++ src/main/java/com/project/backend/domain/event/converter/RecurrenceGroupSpec.java
@@ -1,8 +1,8 @@
 package com.project.backend.domain.event.converter;
 
-import com.project.backend.domain.event.enums.MonthlyType;
-import com.project.backend.domain.event.enums.RecurrenceEndType;
-import com.project.backend.domain.event.enums.RecurrenceFrequency;
+import com.project.backend.domain.common.recurrence.enums.MonthlyType;
+import com.project.backend.domain.common.recurrence.enums.RecurrenceEndType;
+import com.project.backend.domain.common.recurrence.enums.RecurrenceFrequency;
 import lombok.Builder;
 
 import java.time.DayOfWeek;
