+++ src/main/java/com/project/backend/domain/event/entity/Event.java
@@ -1,7 +1,7 @@
 package com.project.backend.domain.event.entity;
 
 import com.project.backend.domain.event.enums.EventColor;
-import com.project.backend.domain.event.enums.RecurrenceFrequency;
+import com.project.backend.domain.common.recurrence.enums.RecurrenceFrequency;
 import com.project.backend.global.entity.BaseEntity;
 import com.project.backend.domain.member.entity.Member;
 import jakarta.persistence.*;
