+++ src/main/java/com/project/backend/domain/event/entity/RecurrenceException.java
@@ -1,8 +1,8 @@
 package com.project.backend.domain.event.entity;
 
 import com.project.backend.domain.event.enums.EventColor;
-import com.project.backend.domain.event.enums.ExceptionType;
-import com.project.backend.domain.event.enums.RecurrenceFrequency;
+import com.project.backend.domain.common.recurrence.enums.ExceptionType;
+import com.project.backend.domain.common.recurrence.enums.RecurrenceFrequency;
 import com.project.backend.global.entity.BaseEntity;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -10,7 +10,6 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.Objects;
