+++ src/main/java/com/project/backend/domain/suggestion/generation/candidate/RecurrenceSuggestionCandidate.java
@@ -2,9 +2,9 @@ package com.project.backend.domain.suggestion.generation.candidate;
 
 import com.project.backend.domain.event.entity.Event;
 import com.project.backend.domain.event.entity.RecurrenceGroup;
-import com.project.backend.domain.event.enums.MonthlyType;
-import com.project.backend.domain.event.enums.RecurrenceEndType;
-import com.project.backend.domain.event.enums.RecurrenceFrequency;
+import com.project.backend.domain.common.recurrence.enums.MonthlyType;
+import com.project.backend.domain.common.recurrence.enums.RecurrenceEndType;
+import com.project.backend.domain.common.recurrence.enums.RecurrenceFrequency;
 import com.project.backend.domain.member.entity.Member;
 import com.project.backend.domain.suggestion.enums.Category;
 import com.project.backend.domain.todo.entity.Todo;
