+++ src/main/java/com/project/backend/domain/suggestion/invalidation/fingerprint/RecurrenceGroupFingerPrint.java
@@ -1,8 +1,10 @@
 package com.project.backend.domain.suggestion.invalidation.fingerprint;
 
+import com.project.backend.domain.common.recurrence.enums.MonthlyType;
 import com.project.backend.domain.common.recurrence.enums.MonthlyWeekdayRule;
+import com.project.backend.domain.common.recurrence.enums.RecurrenceEndType;
+import com.project.backend.domain.common.recurrence.enums.RecurrenceFrequency;
 import com.project.backend.domain.event.entity.RecurrenceGroup;
-import com.project.backend.domain.event.enums.*;
 import com.project.backend.global.recurrence.util.RecurrenceUtils;
 
 import java.time.LocalDate;
