+++ src/main/java/com/project/backend/domain/suggestion/invalidation/fingerprint/TodoRecurrenceGroupFingerPrint.java
@@ -1,8 +1,8 @@
 package com.project.backend.domain.suggestion.invalidation.fingerprint;
 
-import com.project.backend.domain.event.enums.MonthlyType;
-import com.project.backend.domain.event.enums.RecurrenceEndType;
-import com.project.backend.domain.event.enums.RecurrenceFrequency;
+import com.project.backend.domain.common.recurrence.enums.MonthlyType;
+import com.project.backend.domain.common.recurrence.enums.RecurrenceEndType;
+import com.project.backend.domain.common.recurrence.enums.RecurrenceFrequency;
 import com.project.backend.domain.todo.entity.TodoRecurrenceGroup;
 
 import java.time.LocalDate;
